Hertz Bug report fix SCRAPE-146:
It is found that with User_Agent is passed via requests (and hence Robo browser) that Hertz page fails to respond and times out.  This seems to be a combination of Requests and Hertz.  It cannot be simulated in postman which works with same header!
The fix sets the the header to None so it will not be sent when requesting the statement. The login URL is not adversely affected by the user agent but other urls appear to be.

(released on hotfix/1.9.4-SCRAPE-146-Hertz-Agent branch
 ie See Jira SCRAPE-146 for issues and comments)